kanban_api/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2-binary
requests
drf_yasg
